+++ src/bookdown/commands/ServeBook.ts
@@ -5,14 +5,14 @@ import { BaseCommand } from "../../common/BaseCommand";
 import { verbatimOutput } from "../../utils/verbatimOutput";
 import { spawn } from "child_process";
 
-// __title__ = "Blogdown: Serve Site";
-export class ServeSite extends BaseCommand {
+// __title__ = "Bookdown: Serve Book";
+export class ServeBook extends BaseCommand {
   private _outputChannel: vscode.OutputChannel | undefined;
   init() {
-    this._outputChannel = vscode.window.createOutputChannel("Blogdown");
+    this._outputChannel = vscode.window.createOutputChannel("Bookdown");
   }
   async run() {
-    let p = spawn(Rscript(Rcall("blogdown::serve_site")), { cwd: vscode.workspace.workspaceFolders![0].uri.path, shell: true });
+    let p = spawn(Rscript(Rcall("bookdown::serve_book()")), { cwd: vscode.workspace.workspaceFolders![0].uri.path, shell: true });
     console.log(p);
     verbatimOutput(p, this._outputChannel as vscode.OutputChannel);
   }